requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
aptdaemon==2.0.2
attrs==25.1.0
autocommand==2.2.2
babel==2.17.0
bcc==0.30.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brlapi==0.8.6
Brotli==1.1.0
build==1.2.2
CacheControl==0.14.2
certifi==2025.1.31
chardet==5.2.0
cleo==2.2.1
click==8.1.8
cloud-init==25.2
command-not-found==0.3
configobj==5.0.9
contourpy==1.3.1
cryptography==43.0.0
cssselect==1.3.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
distro-info==1.13
dnspython==2.7.0
docker==7.1.0
dulwich==0.22.7
email_validator==2.2.0
fastapi==0.115.11
fastimport==0.9.14
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.2
fonttools==4.55.3
fs==2.4.16
Glances==4.3.0.8
gyp-next==0.16.2
h11==0.14.0
html5lib-modern==1.2
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflect==7.3.1
influxdb==5.3.2
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
keyring==25.6.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.32.0
lxml==5.3.2
lz4==4.4.0+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.0.3
netaddr==1.3.0
netifaces==0.11.0
numpy==2.2.3
oauthlib==3.2.2
olefile==0.47
orjson==3.10.7
packaging==24.2
passlib==1.7.4
pbs-installer==2025.3.11
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.6
ply==3.11
poetry==2.1.1
poetry-core==2.1.1
pooch==1.8.2
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.1
pyasyncore==1.0.2
pycairo==1.27.0
pycryptodomex==3.20.0
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.10.1
pyparsing==3.1.2
pyproject_hooks==1.2.0
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pystache==0.6.5
python-apt==3.0.0
python-dateutil==2.9.0
python-debian==1.0.1+ubuntu1
python-multipart==0.0.20
pytz==2025.1
pyxdg==0.28
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
scipy==1.14.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
sympy==1.13.3
systemd-python==235
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typeguard==4.4.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.17.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.29.1
wadllib==2.0.0
webencodings==0.5.1
wheel==0.45.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==3.21.0
zstandard==0.23.0
